Aula 18 - Adaptação no Caso de uso de deletar pergunta para verificar autor

+++ src/domain/forum/application/use-cases/delete-question.ts
@@ -2,17 +2,24 @@ import QuestionsRepository from '@/domain/forum/application/repositories/questio
 import UniqueEntityId from '@/core/entities/unique-entity-id'
 
 type DeleteQuestionInput = {
+  authorId: UniqueEntityId
   questionId: UniqueEntityId
 }
 
 export default class DeleteQuestion {
   constructor(private questionsRepository: QuestionsRepository) {}
 
-  async execute({ questionId }: DeleteQuestionInput): Promise<void> {
+  async execute({ authorId, questionId }: DeleteQuestionInput): Promise<void> {
     const question = await this.questionsRepository.findById(questionId)
 
     if (!question) {
       throw new Error(`A pergunta com o id: ${questionId} não foi encontrada`)
+    }
+
+    if (authorId !== question.authorId) {
+      throw new Error(
+        'Não é permitido deleter a pergunta de um usuário diferente',
+      )
     }
 
     await this.questionsRepository.delete(questionId)
